backend/requirements.txt
# ANCHOR Backend Dependencies
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Security & Cryptography
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
pynacl>=1.5.0
webauthn>=2.0.0
cryptography>=41.0.0

# Utilities
httpx>=0.26.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

